Move color classes into Xabbo.Scripter.Appearance namespace
